Add files via upload
The first map is a choropleth map that displays the distribution across the territory of the Province of Santa Fe, showing the number of people injured by firearms in the year 2022. The second map is a point map within the Department of Rosario, representing the territorial distribution of people injured by firearms. Then, there's a point map of the city of Rosario, followed by a grid map illustrating territorial concentrations of people injured by firearms. Similar maps were created for the Department of La Capital and the city of Santa Fe.